+++ config/autoload.php
@@ -24,29 +24,30 @@ ClassLoader::addNamespaces(array
 ClassLoader::addClasses(array
 (
 	// Classes
-	'Tastaturberuf\AnyStores'               => 'system/modules/anyStores/classes/AnyStores.php',
-	'Tastaturberuf\AnyStoresDcaHelper'      => 'system/modules/anyStores/classes/AnyStoresDcaHelper.php',
-	'Tastaturberuf\AnyStoresHooks'          => 'system/modules/anyStores/classes/AnyStoresHooks.php',
-	'Tastaturberuf\GoogleMaps'              => 'system/modules/anyStores/classes/API/GoogleMaps.php',
-	'Tastaturberuf\OpenStreetMap'           => 'system/modules/anyStores/classes/API/OpenStreetMap.php',
+	'Tastaturberuf\AnyStores'                => 'system/modules/anyStores/classes/AnyStores.php',
+	'Tastaturberuf\AnyStoresDcaHelper'       => 'system/modules/anyStores/classes/AnyStoresDcaHelper.php',
+	'Tastaturberuf\AnyStoresHooks'           => 'system/modules/anyStores/classes/AnyStoresHooks.php',
+	'Tastaturberuf\GoogleMaps'               => 'system/modules/anyStores/classes/API/GoogleMaps.php',
+	'Tastaturberuf\OpenStreetMap'            => 'system/modules/anyStores/classes/API/OpenStreetMap.php',
 
 	// Controllers
-	'Tastaturberuf\AnyStoresRunonce'        => 'system/modules/anyStores/controllers/AnyStoresRunonce.php',
-	'Tastaturberuf\FrontendAjax'            => 'system/modules/anyStores/controllers/FrontendAjax.php',
+	'Tastaturberuf\AnyStoresRunonce'         => 'system/modules/anyStores/controllers/AnyStoresRunonce.php',
+	'Tastaturberuf\FrontendAjax'             => 'system/modules/anyStores/controllers/FrontendAjax.php',
 
 	// Forms
-	'Tastaturberuf\FormStores'              => 'system/modules/anyStores/forms/FormStores.php',
+	'Tastaturberuf\FormStores'               => 'system/modules/anyStores/forms/FormStores.php',
 
 	// Models
-	'Tastaturberuf\AnyStoresCategoryModel'  => 'system/modules/anyStores/models/AnyStoresCategoryModel.php',
-	'Tastaturberuf\AnyStoresModel'          => 'system/modules/anyStores/models/AnyStoresModel.php',
+	'Tastaturberuf\AnyStoresCategoryModel'   => 'system/modules/anyStores/models/AnyStoresCategoryModel.php',
+	'Tastaturberuf\AnyStoresModel'           => 'system/modules/anyStores/models/AnyStoresModel.php',
 
 	// Modules
-	'Tastaturberuf\ModuleAnyStoresDetails'  => 'system/modules/anyStores/modules/ModuleAnyStoresDetails.php',
-	'Tastaturberuf\ModuleAnyStoresImporter' => 'system/modules/anyStores/modules/ModuleAnyStoresImporter.php',
-	'Tastaturberuf\ModuleAnyStoresList'     => 'system/modules/anyStores/modules/ModuleAnyStoresList.php',
-	'Tastaturberuf\ModuleAnyStoresMap'      => 'system/modules/anyStores/modules/ModuleAnyStoresMap.php',
-	'Tastaturberuf\ModuleAnyStoresSearch'   => 'system/modules/anyStores/modules/ModuleAnyStoresSearch.php',
+	'Tastaturberuf\ModuleAnyStoresDetails'   => 'system/modules/anyStores/modules/ModuleAnyStoresDetails.php',
+	'Tastaturberuf\ModuleAnyStoresImporter'  => 'system/modules/anyStores/modules/ModuleAnyStoresImporter.php',
+	'Tastaturberuf\ModuleAnyStoresList'      => 'system/modules/anyStores/modules/ModuleAnyStoresList.php',
+	'Tastaturberuf\ModuleAnyStoresMap'       => 'system/modules/anyStores/modules/ModuleAnyStoresMap.php',
+	'Tastaturberuf\ModuleAnyStoresSearch'    => 'system/modules/anyStores/modules/ModuleAnyStoresSearch.php',
+	'Tastaturberuf\ModuleAnyStoresSearchMap' => 'system/modules/anyStores/modules/ModuleAnyStoresSearchMap.php',
 ));
 
 
@@ -63,4 +64,5 @@ TemplateLoader::addFiles(array
 	'mod_anystores_list'      => 'system/modules/anyStores/templates/modules',
 	'mod_anystores_map'       => 'system/modules/anyStores/templates/modules',
 	'mod_anystores_search'    => 'system/modules/anyStores/templates/modules',
+	'mod_anystores_searchmap' => 'system/modules/anyStores/templates/modules',
 ));