+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   IMPORTANT: To override templates put it in the templates root folder! The AJAX request has no layout object and can not find the template in other folders.
 - runonce.php support
 - add full user / groups permission support
+- add search map module. You can see the results on a map
 
 ###Changed
 - change ajax request to new class FrontendAjax.php